refactor: updating to new architecture
