+++ package.json
@@ -12,6 +12,7 @@
     "i18next-browser-languagedetector": "^6.1.4",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
+    "react-icons": "^4.4.0",
     "styled-components": "^5.3.5",
     "vite-tsconfig-paths": "^3.5.0"
   },
