+++ src/components/index.ts
@@ -1,3 +1,4 @@
-export { default as Header } from "./Header";
-export { Post } from "./Post";
 export { default as Button } from "./Button";
+export { default as Comment } from "./Comment";
+export { default as Header } from "./Header";
+export { default as Post } from "./Post";
