+++ src/translate/locales/en-us/index.ts
@@ -1,5 +1,14 @@
 export default {
   translations: {
     publish: "Publish",
+    publishedAt: "Published",
+    times: {
+      now: "now",
+      minutes: "há alguns minutos",
+      hours: "há algumas horas",
+      days: "há alguns dias",
+      weeks: "há algumas semanas",
+      months: "há alguns meses",
+    },
   },
 };
